Update ProjectMembership to include 'club'
This change makes sure that for each ProjectMembership the club, for
which a user is working, is known. This also makes it easy to know
the ProjectMemberships that need to be deleted when a Club is removed
from a Project. This commit also adds required validation to the
corresponding serializer.