Update db schema for Project model
This commit updates the schema for the Project model to include the
owner club as a foreign key. This commit also makes the required changes
to the methods provided by this model and the permissions, views,
serializers and filters. The requried database migrations are also
pushed as part of this commit.